+++ boa_examples/Cargo.toml
@@ -13,5 +13,6 @@ rust-version.workspace = true
 
 [dependencies]
 boa_engine = { workspace = true, features = ["console"] }
+boa_interner = { workspace = true }
 boa_gc.workspace = true
 gc = "0.4.1"
